22_Data_clean_n_remove_deduplication/requirements.txt
markitdown[all]
trafilatura
playwright
datasketch
python-docx
openpyxl
pandas
python-pptx
langchain_community
langchain_ollama
ragas
datasets
pypdf
rapidfuzz
markdown
newspaper4k
requests
urllib3
charset_normalizer

